+++ edx_lint/cmd/list.py
@@ -6,7 +6,7 @@ from edx_lint.metadata import KNOWN_FILES
 def list_main(argv_unused):  # pylint: disable=unused-argument
     """
     list
-        List the FILENAMEs that edx_lint can provide.
+        List the files that edx_lint can provide.
     """
     print("edx_lint knows about these files:")
     for filename, metadata in KNOWN_FILES.items():